refs #140: migrate autoload to psr-4
