+++ composer.json
@@ -8,7 +8,7 @@
         "ext-json": "*"
     },
     "require-dev": {
-        "nette/tester": "~1.3",
+        "nette/tester": "~1.7",
         "jakub-onderka/php-console-highlighter": "~0.3",
         "squizlabs/php_codesniffer": "~2.7"
     },
@@ -22,9 +22,11 @@
         }
     ],
     "autoload": {
-        "classmap": [
-            "./"
-        ]
+        "psr-4": {
+            "JakubOnderka\\PhpParallelLint\\": "src/"
+        }
     },
-    "bin": ["parallel-lint"]
+    "bin": [
+        "parallel-lint"
+    ]
 }
