+++ src/ArgumentsArrayIterator.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace JakubOnderka\PhpParallelLint;
+
+class ArgumentsArrayIterator extends \ArrayIterator
+{
+    public function getNext()
+    {
+        $this->next();
+
+        return $this->current();
+    }
+}
