+++ src/ConsoleWriter.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace JakubOnderka\PhpParallelLint;
+
+class ConsoleWriter implements IWriter
+{
+    /**
+     * @param string $string
+     */
+    public function write($string)
+    {
+        echo $string;
+    }
+}
