+++ src/FileWriter.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace JakubOnderka\PhpParallelLint;
+
+class FileWriter implements IWriter
+{
+    /** @var string */
+    protected $logFile;
+
+    /** @var string */
+    protected $buffer;
+
+    public function __construct($logFile)
+    {
+        $this->logFile = $logFile;
+    }
+
+    public function write($string)
+    {
+        $this->buffer .= $string;
+    }
+
+    public function __destruct()
+    {
+        file_put_contents($this->logFile, $this->buffer);
+    }
+}
