+++ src/Manager.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace JakubOnderka\PhpParallelLint;
 
 use JakubOnderka\PhpParallelLint\Process\GitBlameProcess;
@@ -11,6 +12,7 @@ class Manager
 
     /**
      * @param null|Settings $settings
+     *
      * @return Result
      * @throws Exception
      * @throws \Exception
@@ -72,6 +74,7 @@ class Manager
 
     /**
      * @param Settings $settings
+     *
      * @return Output
      */
     protected function getDefaultOutput(Settings $settings)
@@ -98,6 +101,7 @@ class Manager
     /**
      * @param Result $result
      * @param Settings $settings
+     *
      * @throws Exception
      */
     protected function gitBlame(Result $result, Settings $settings)
@@ -129,22 +133,25 @@ class Manager
      * @param array $paths
      * @param array $extensions
      * @param array $excluded
+     *
      * @return array
      * @throws NotExistsPathException
      */
-    protected function getFilesFromPaths(array $paths, array $extensions, array $excluded = array())
+    protected function getFilesFromPaths(array $paths, array $extensions, array $excluded = [])
     {
         $extensions = array_flip($extensions);
-        $files = array();
+        $files = [];
 
         foreach ($paths as $path) {
             if (is_file($path)) {
                 $files[] = $path;
             } else if (is_dir($path)) {
                 $iterator = new \RecursiveDirectoryIterator($path, \FilesystemIterator::SKIP_DOTS);
+
                 if (!empty($excluded)) {
                     $iterator = new RecursiveDirectoryFilterIterator($iterator, $excluded);
                 }
+
                 $iterator = new \RecursiveIteratorIterator(
                     $iterator,
                     \RecursiveIteratorIterator::LEAVES_ONLY,
@@ -165,93 +172,5 @@ class Manager
         $files = array_unique($files);
 
         return $files;
-    }
-}
-
-class RecursiveDirectoryFilterIterator extends \RecursiveFilterIterator
-{
-    /** @var \RecursiveDirectoryIterator */
-    private $iterator;
-
-    /** @var array */
-    private $excluded = array();
-
-    /**
-     * @param \RecursiveDirectoryIterator $iterator
-     * @param array $excluded
-     */
-    public function __construct(\RecursiveDirectoryIterator $iterator, array $excluded)
-    {
-        parent::__construct($iterator);
-        $this->iterator = $iterator;
-        $this->excluded = array_map(array($this, 'getPathname'), $excluded);
-    }
-
-    /**
-     * (PHP 5 &gt;= 5.1.0)<br/>
-     * Check whether the current element of the iterator is acceptable
-     *
-     * @link http://php.net/manual/en/filteriterator.accept.php
-     * @return bool true if the current element is acceptable, otherwise false.
-     */
-    public function accept()
-    {
-        $current = $this->current()->getPathname();
-        $current = $this->normalizeDirectorySeparator($current);
-
-        if ('.' . DIRECTORY_SEPARATOR !== $current[0] . $current[1]) {
-            $current = '.' . DIRECTORY_SEPARATOR . $current;
-        }
-
-        return !in_array($current, $this->excluded);
-    }
-
-    /**
-     * (PHP 5 &gt;= 5.1.0)<br/>
-     * Check whether the inner iterator's current element has children
-     *
-     * @link http://php.net/manual/en/recursivefilteriterator.haschildren.php
-     * @return bool true if the inner iterator has children, otherwise false
-     */
-    public function hasChildren()
-    {
-        return $this->iterator->hasChildren();
-    }
-
-    /**
-     * (PHP 5 &gt;= 5.1.0)<br/>
-     * Return the inner iterator's children contained in a RecursiveFilterIterator
-     *
-     * @link http://php.net/manual/en/recursivefilteriterator.getchildren.php
-     * @return \RecursiveFilterIterator containing the inner iterator's children.
-     */
-    public function getChildren()
-    {
-        return new self($this->iterator->getChildren(), $this->excluded);
-    }
-
-    /**
-     * @param string $file
-     * @return string
-     */
-    private function getPathname($file)
-    {
-        $file = $this->normalizeDirectorySeparator($file);
-
-        if ('.' . DIRECTORY_SEPARATOR !== $file[0] . $file[1]) {
-            $file = '.' . DIRECTORY_SEPARATOR . $file;
-        }
-
-        $directoryFile = new \SplFileInfo($file);
-        return $directoryFile->getPathname();
-    }
-
-    /**
-     * @param string $file
-     * @return string
-     */
-    private function normalizeDirectorySeparator($file)
-    {
-        return str_replace(array('\\', '/'), DIRECTORY_SEPARATOR, $file);
     }
 }
