+++ src/NullWriter.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace JakubOnderka\PhpParallelLint;
+
+class NullWriter implements IWriter
+{
+    /**
+     * @param string $string
+     */
+    public function write($string)
+    {
+
+    }
+}
