+++ src/RunTimeException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace JakubOnderka\PhpParallelLint;
+
+class RunTimeException extends Exception
+{
+
+}
+
