+++ src/Settings.php
@@ -1,9 +1,9 @@
 <?php
+
 namespace JakubOnderka\PhpParallelLint;
 
 class Settings
 {
-
     /**
      * constants for enum settings
      */
@@ -17,78 +17,91 @@ class Settings
 
     /**
      * Path to PHP executable
+     *
      * @var string
      */
     public $phpExecutable = 'php';
 
     /**
      * Check code inside PHP opening short tag <? or <?= in PHP 5.3
+     *
      * @var bool
      */
     public $shortTag = false;
 
     /**
      * Check PHP code inside ASP-style <% %> tags.
+     *
      * @var bool
      */
     public $aspTags = false;
 
     /**
      * Number of jobs running in same time
+     *
      * @var int
      */
     public $parallelJobs = 10;
 
     /**
      * If path contains directory, only file with these extensions are checked
+     *
      * @var array
      */
-    public $extensions = array('php', 'phtml', 'php3', 'php4', 'php5', 'phpt');
+    public $extensions = ['php', 'phtml', 'php3', 'php4', 'php5', 'phpt'];
 
     /**
      * Array of file or directories to check
+     *
      * @var array
      */
-    public $paths = array();
+    public $paths = [];
 
     /**
      * Don't check files or directories
+     *
      * @var array
      */
-    public $excluded = array();
+    public $excluded = [];
 
     /**
      * Mode for color detection. Possible values: self::FORCED, self::DISABLED and self::AUTODETECT
+     *
      * @var string
      */
     public $colors = self::AUTODETECT;
 
     /**
      * Show progress in text output
+     *
      * @var bool
      */
     public $showProgress = true;
 
     /**
      * Output format (see FORMAT_* constants)
+     *
      * @var string
      */
     public $format = self::FORMAT_TEXT;
 
     /**
      * Read files and folder to tests from standard input (blocking)
+     *
      * @var bool
      */
     public $stdin = false;
 
     /**
      * Try to show git blame for row with error
+     *
      * @var bool
      */
     public $blame = false;
 
     /**
      * Path to git executable for blame
+     *
      * @var string
      */
     public $gitExecutable = 'git';
@@ -113,12 +126,13 @@ class Settings
 
     /**
      * @param array $arguments
+     *
      * @return Settings
      * @throws InvalidArgumentException
      */
     public static function parseArguments(array $arguments)
     {
-        $arguments = new ArrayIterator(array_slice($arguments, 1));
+        $arguments = new ArgumentsArrayIterator(array_slice($arguments, 1));
         $settings = new self;
 
         // Use the currently invoked php as the default if possible
@@ -214,19 +228,11 @@ class Settings
         $content = stream_get_contents(STDIN);
 
         if (empty($content)) {
-            return array();
+            return [];
         }
 
         $lines = explode("\n", rtrim($content));
-        return array_map('rtrim', $lines);
-    }
-}
 
-class ArrayIterator extends \ArrayIterator
-{
-    public function getNext()
-    {
-        $this->next();
-        return $this->current();
+        return array_map('rtrim', $lines);
     }
 }
